+++ src/hooked/hooked.hpp
@@ -3,5 +3,6 @@
 #include "pause_layer.hpp"
 #include "gj_base_game_layer.hpp"
 #include "level_info_layer.hpp"
+#include "loading_layer.hpp"
 
 #include "error_check_layer_hooks.hpp"